Fix for current devel, removes `untyped` return type for iterators (#55)
* removes `untyped` return value for iterators,fix nameless tuple test

Not supported on current devel anymore. The `items` iterator for
tuples comes with a caveat. In case it's used on a tuple containing
different types, e.g. `tuple[a: float, b: int]` it will error at
compile time, because `auto` sets the return value of the iterator to
the first type it encounters.

Since items seems to be used only to flatten nested collections
anyways, this is not an issue, since then we demand the same type
anyways.

Also fixes a test for nameless tuples. Rendering of nameless tuples it
seems has changed recently, so that a nameless tuple is printed
without the `Field0` etc. names.

* add `a[int] is T` to two concepts to remove some "XDeclaredNotUsed"

* add note about `items` constraint